refactor: rename parameter to apiKey
In a call a spade a spade decision, the `parameter` authentication
method was renamed to `apiKey` to decrease the cognitive load and fix
issues where the terminology clashes like in #5369.

Fixes #5369
