+++ app/connector/rest-swagger/src/main/resources/META-INF/syndesis/connector/rest-swagger.json
@@ -99,8 +99,8 @@
           "value": "oauth2"
         },
         {
-          "label": "parameter",
-          "value": "parameter"
+          "label": "apiKey",
+          "value": "apiKey"
         }
       ],
       "javaType": "java.lang.String",
